Move RawValue and read_value to scylla-cql-core
I duplicated read_raw_bytes, because I don't want to make it public in
core, but it is used in many places in scylla-cql.
